+++ CHANGELOG.md
@@ -16,6 +16,7 @@
   * Renaming _execPrePrepareUndoRedoInfo and _execPreCreateUndoRedoCommand at MPCommand.
   * Creating MPAddLineSegmentCommand.atExistingLineSegmentPosition().
   * Removing all references to web version.
+  * Making custom exceptions be presented by showUnhandledErrorDialog() at main.dart.
 
 ## 0.2.35 - 2026-02-13 - The [Chiquita Bacana's daughter](https://www.youtube.com/watch?v=QajgvvuFOSg) release
 * New features: